Apply Anti aliasing diff
